0.4.13

+++ Tx.py
@@ -11,7 +11,6 @@ import serial
 import subprocess
 import sys
 import time
-from time import sleep
 
 
 
@@ -20,7 +19,7 @@ from time import sleep
 ######################################################################
 
 # TFC-CEV (Cascading Encryption Version) || Tx.py
-version = 'CEV 0.4.12 beta'
+version = 'CEV 0.4.13 beta'
 
 """
 This software is part of the TFC application, which is free software:
@@ -49,7 +48,7 @@ lMsgSleep          = 0.2
 debugging          = False
 emergencyExit      = False
 randomSleep        = False
-localTesting       = False
+localTesting       = True
 
 if not localTesting:
     port = serial.Serial('/dev/ttyAMA0', baudrate=9600, timeout=0.1)
@@ -971,7 +970,6 @@ def get_keyset(xmpp, output=True):
 
 
 
-
 def rotate_keyset(xmpp):
     try:
 
@@ -1001,7 +999,6 @@ def rotate_keyset(xmpp):
 
     except IOError:
         exit_with_msg('CRITICAL ERROR! Keyfile ' + xmpp + '.e\ncould not be loaded.')
-
 
 
 
@@ -1241,7 +1238,8 @@ def get_autotabs(contacts):
     # Add commands to autoTabs.
     cList = ['about' , 'add '  , 'clear', 'create ' , 'exit',
              'file ' , 'group ', 'help' , 'logging ', 'msg ',
-             'newkf ', 'nick ' , 'quit' , 'rm '     , 'select ']
+             'newkf ', 'nick ' , 'quit' , 'rm '     , 'select ',
+             'store ']
 
     for command in cList:
         autoTabs.append(command)
@@ -1270,14 +1268,13 @@ def search_keyfiles():
 
 def chk_pkgSize():
     if PkgSize > 250:
-        exit_with_msg('ERROR! Maximum length of packet is 250 characters.\n' \
+        exit_with_msg('ERROR! Maximum length of packet is 250 characters.\n'
               'Please fix the value \'PkgSize\'and restart TFC.')
 
 
     if PkgSize < 25:
         exit_with_msg('ERROR! Minimum length of packet is 25 characters.'
                       '\nPlease fix the value \'PkgSize\' and restart TFC.')
-
 
 
 
@@ -1314,24 +1311,49 @@ def padding(string):
 
 def long_msg_preprocess(string):
 
-    string = string.strip('\n')
-    msgA   = [string[i:i + (PkgSize - 2)] for i in range(0, len(string), (PkgSize - 2) )] # Each packet is left one char shorter than padding to prevent dummy blocks.
+    preprocessFile = False
+    preprocessMsg  = False
 
-    for i in xrange( len(msgA) ):          # Add 'a' in front of every packet.
-        msgA[i] = 'a' + msgA[i]            # 'a' tells Rx.py the message should be appended to long message.
+    # Strip leading white space.
+    while string.startswith(' '):
+        string = string[1:]
 
-                                           # Replace 'a' with 'e' in last packet.
-    msgA[-1] = 'e' + msgA[-1][1:]          # 'e' tells Rx.py the message is last one of long message, so Rx.py knows it must show the entire long message.
+    # Remove 'm' from start of string.
+    if string.startswith('m'):
+        preprocessMsg = True
+    if string.startswith('f'):
+        preprocessFile = True
 
-                                           # Replace 'a' with 'l' in first packet.
-    msgA[0]  = 'l' + msgA[0][1:]           # 'l' tells Rx.py the message extends over multiple packets.
+    # Remove new lines and 'm' or 'f' header.
+    string  = string[1:].strip('\n')
+
+    # Append cryptographic hash of message / file to packet.
+    string  = string + keccak_256(string)
+
+    # Split packet to list the items of which are 3 char shorter than max length:
+    # This leaves room for 2 char header and prevents dummy blocks when padding.
+    packetA = [string[i:i + (PkgSize - 3)] for i in range(0, len(string), (PkgSize - 3))]
+
+
+    if preprocessMsg:                        # All packets to recipients have header {s,l,a,e}{m,f}.
+        for i in xrange(len(packetA)):
+            packetA[i] = 'am' + packetA[i]   # 'am' stands for appended message packets.
+        packetA[-1] = 'em' + packetA[-1][2:] # 'em' stands for end message packets.
+        packetA[0]  = 'lm' + packetA[0][2:]  # 'lm' stands for start of long message packets.
+
+    if preprocessFile:
+        for i in xrange(len(packetA)):
+            packetA[i] = 'af' + packetA[i]   # 'af' stands for appended file packets.
+        packetA[-1] = 'ef' + packetA[-1][2:] # 'ef' stands for end file packets.
+        packetA[0]  = 'lf' + packetA[0][2:]  # 'lf' stands for start of long file packets.
+
 
     if debugging:
-        print 'M(long_msg_preprocess): Processed long message to following packets:'
-        for item in msgA:
+        print '\nM(long_msg_preprocess): Processed long message to following packets:'
+        for item in packetA:
             print '"' + item + '"'
 
-    return msgA
+    return packetA
 
 
 
@@ -1339,18 +1361,12 @@ def long_msg_process(userInput, xmpp):
 
     packetList = long_msg_preprocess(userInput)
 
-    print '\nTransferring message in ' + str(len(packetList)) + ' parts. ^C cancels'
+    if userInput.startswith('f'):
+        print '\nTransferring file over ' + str(len(packetList)) + ' packets. ^C cancels'
+    if userInput.startswith('m'):
+        print '\nTransferring message over ' + str(len(packetList)) + ' packets. ^C cancels'
 
-    halt = False
     for packet in packetList:
-
-        if halt:
-            os.system('clear')
-            if userInput.startswith('TFCFILE'):
-                print '\nFile transmission interrupted by user.\n'
-            else:
-                print '\nMessage transmission interrupted by user.\n'
-            return None
 
         try:
             keyID     = get_keyID(xmpp)
@@ -1367,21 +1383,47 @@ def long_msg_process(userInput, xmpp):
             if randomSleep:
                 sleepTime = random.uniform(0, maxSleepTime)
                 print 'Sleeping ' + str(sleepTime) + ' seconds to obfuscate long message.'
-                sleep(sleepTime)
+                time.sleep(sleepTime)
 
             # Minimum sleep time ensures XMPP server is not flooded.
-            sleep(lMsgSleep)
+            time.sleep(lMsgSleep)
 
         except KeyboardInterrupt:
-            halt = True
+
+            # If user interrupts packet transmission, send encrypted notification.
+            os.system('clear')
+            if userInput.startswith('f'):
+                print '\nFile transmission interrupted by user.\n'
+                cancelMsg = 'cf'
+            else:
+                print '\nMessage transmission interrupted by user.\n'
+                cancelMsg = 'cm'
+
+            keyID     = get_keyID(xmpp)
+
+            paddedMsg = padding(cancelMsg)
+            ctWithTag = encrypt(xmpp, paddedMsg)
+
+            encoded   = b64e(ctWithTag)
+            checksum  = crc32(encoded + '|' + str(keyID))
+
+            write_keyID(xmpp, keyID + 1)
+            output_message(xmpp, encoded, keyID, checksum)
+
+            return None
+
+    if userInput.startswith('f'):
+        print '\nFile transmission complete.\n'
+    if userInput.startswith('m'):
+        print '\nMessage transmission complete.\n'
 
 
 
 def short_msg_process(plaintext, xmpp):
 
     keyID     = get_keyID(xmpp)
-
-    paddedMsg = padding('s' + plaintext) # 's' tells Rx.py the message is only one packet long.
+                                         # All packets to recipients have header {s,l,a,e}{m,f}.
+    paddedMsg = padding('s' + plaintext) # 's' stands for single packet messages.
     ctWithTag = encrypt(xmpp, paddedMsg)
 
     encoded  = b64e(ctWithTag)
@@ -1407,15 +1449,17 @@ def cmd_msg_process(command):
 
 
 
-def quit_process(output = False):
+def quit_process(output=False):
     os.system('clear')
+
     if output:
         print '\nExiting TFC\n'
+
     if localTesting:
         with open('TxOutput', 'w+') as file:
-            file.write('exitTFC\n')
+            file.write('EXITTFC\n')
     else:
-        port.write('exitTFC\n')
+        port.write('EXITTFC\n')
     exit()
 
 
@@ -1481,12 +1525,14 @@ def print_help():
         print '/file <filename>\nSend file to recipient'                      + le
         print '/nick <nick>\nChange contact nickname'                         + le
         print '/quit & /exit\nExit TFC'                                       + le
+        print '/store <b64 file> <output file>\nDecodes received file'        + le
         print '/group\nList group members'                                    + le
         print '/groups\nList available groups'                                + le
         print '/group <groupname>\nSelect group'                              + le
         print '/group create <groupname> <xmpp1> <xmpp2>\nCreate new group'   + le
         print '/group add <groupname> <xmpp1> <xmpp2>\nAdd xmpp to group'     + le
         print '/group rm <groupname> <xmpp1> <xmpp2>\nRemove xmpp from group' + le
+        print '/store <tmp f.name> <output f.name>\nSave received tmp file'   + le
         print '/shift + PgUp/PgDn\nScroll terminal up/dn'                     + le
         print ''
 
@@ -1506,13 +1552,16 @@ def print_help():
         print       ' /paste'                                      + 16 * ' ' + 'Enable paste-mode'
         print       ' /nick <nick>'                                + 10 * ' ' + 'Change contact\'s nickname on Tx.py & Rx.py'
         print       ' /quit & /exit'                               + 9  * ' ' + 'Exits TFC'
+        print                                                      ttyW * '-'
         print       ' /group'                                      + 16 * ' ' + 'List group members'
         print       ' /groups'                                     + 15 * ' ' + 'List available groups'
         print       ' /group <groupname>'                          + 4  * ' ' + 'Select group\n'
-        print       ' /group create <groupname> <xmpp1> <xmpp2>'   + 5  * ' ' + '\n Create new group called <groupname>, add list of xmpp-addresses.\n'
-        print       ' /group add <groupname> <xmpp1> <xmpp2>'      + 8  * ' ' + '\n Add xmpp-addresses to group <groupname>\n'
-        print       ' /group rm <groupname> <xmpp1> <xmpp2>'       + 8  * ' ' + '\n Remove xmpp-addresses from group <groupname>\n'
-        print       ' shift + PgUp/PgDn'                           + 5  * ' ' + 'Scroll terminal up/down\n'
+        print       ' /group create <groupname> <xmpp1> <xmpp2>'   + 1  * ' ' + '\n Create new group called <groupname>, add xmpp-addresses.\n'
+        print       ' /group add <groupname> <xmpp1> <xmpp2>'      + 1  * ' ' + '\n Add xmpp-addresses to group <groupname>\n'
+        print       ' /group rm <groupname> <xmpp1> <xmpp2>'       + 1  * ' ' + '\n Remove xmpp-addresses from group <groupname>'
+        print                                                      ttyW * '-'
+        print       ' /store <b64 file name> <file name>'          + 1  * ' ' + '\n Decodes received tmp file and stores it as <file name> \n'
+        print       ' shift + PgUp/PgDn'                           + 1  * ' ' + '\n Scroll terminal up/down\n'
 
 
 
@@ -1546,7 +1595,7 @@ def select_contact(idSelDist='', contactNo='', menu=True):
         try:
             # If no parameter about contact selection is passed to function, ask for input
             if contactNo == '':
-                selection = (raw_input('\nSelect contact:' + (idSelDist - len('Select contact:') ) * ' ') )
+                selection = (raw_input('\nSelect contact:' + (idSelDist - len('Select contact:')) * ' '))
 
                 intSelection = int(selection)
             else:
@@ -1627,7 +1676,7 @@ def exit_with_msg(message, error=True):
 #                          GROUP MANAGEMENT                          #
 ######################################################################
 
-def group_create(groupName, newMembers = []):
+def group_create(groupName, newMembers=[]):
 
     if os.path.isfile('g.' + groupName + '.tfc'):
         if not (raw_input('Group \'' + groupName + '\' already exists. Type YES to overwrite: ') == 'YES'):
@@ -1844,13 +1893,13 @@ while True:
             lines     = sys.stdin.readlines()
             userInput = '\n' + ''.join(lines)
             print '\nSending...'
-            sleep(0.1)
+            time.sleep(0.1)
 
         except KeyboardInterrupt:
             pastemode = False
             userInput = ''
             print '\nClosing paste mode...'
-            sleep(0.4)
+            time.sleep(0.4)
             os.system('clear')
             continue
 
@@ -2038,12 +2087,13 @@ while True:
     if userInput == '  ' or userInput == '/clear':
         if localTesting:
             with open('TxOutput', 'w+') as file:
-                file.write('clearScreen ' + xmpp + '\n')
+                file.write('CLEARSCREEN ' + xmpp + '\n')
         else:
-            port.write('clearScreen ' + xmpp + '\n')
+            port.write('CLEARSCREEN ' + xmpp + '\n')
 
         os.system('clear')
         continue
+
 
     if userInput == '/quit' or userInput == '/exit':
         quit_process(True)
@@ -2129,7 +2179,7 @@ while True:
 
                 selectedGroup = ''
                 os.system('clear')
-                print 'Now sending messages to ' + nick + ' (' + xmpp + ')\n'
+                print 'Now sending messages to ' + nick + ' (' + xmpp[3:] + ')\n'
                 continue
 
         except ValueError:
@@ -2142,11 +2192,11 @@ while True:
 
 
     ##################################################################
-    ##                      COMMANDS TO RX.py                       ##
+    ##                      ENCRYPTED COMMANDS                      ##
     ##################################################################
 
-    if userInput.startswith('/nick ') or userInput.startswith('/logging '):
-        command    = ''
+    if userInput.startswith('/nick ') or userInput.startswith('/logging ') or userInput.startswith('/store ') or userInput.startswith('/newkf '):
+        command = ''
 
         # Check that local keyfile exists.
         if not os.path.isfile('tx.local.e'):
@@ -2172,9 +2222,11 @@ while True:
             if '=' in newNick or '/' in newNick:
                 print '\nError: Nick can not contain reserved characters / or =\n'
                 continue
+
             if (newNick == 'tx.local') or (newNick == 'rx.local'):
                 print '\nError: Can\'t give reserved nick of local.e to contact.\n'
                 continue
+
             if newNick in groupFileList:
                 print '\nError: Nick is in use for group.\n'
                 continue
@@ -2184,7 +2236,7 @@ while True:
 
             os.system('clear')
             print '\nChanged ' + xmpp[3:] + ' nick to ' + nick + '\n'
-            command = 'rx.' + xmpp[3:] + '/nick=' + nick
+            command = 'NICK rx.' + xmpp[3:] + ' ' + nick
 
 
         ###################
@@ -2195,10 +2247,10 @@ while True:
             value = str(userInput.split(' ')[1])
 
             if value   == 'on':
-                command = 'logson'
+                command = 'LOGSON'
 
             elif value == 'off':
-                command = 'logsoff'
+                command = 'LOGSOFF'
 
             else:
                 print '\nError: Invalid command\n'
@@ -2206,12 +2258,34 @@ while True:
 
 
         ########################
+        #     FILE STORING     #
+        ########################
+        if userInput.startswith('/store '):
+            try:
+                params  = userInput.split(' ')
+                command = 'STOREFILE ' + params[1] + ' ' + params[2]
+            except IndexError:
+                print '\nError: Invalid command.\n'
+                continue
+
+
+        ########################
         #    COMMAND PACKET    #
         ########################
-
         cmd_msg_process(command)
-
         continue
+
+
+
+    ################################
+    #     PACKETS TO RECIPIENT     #
+    ################################
+
+    # All packets to recipients have header {s,l,a,e}{m,f}. 'm' stands for message.
+    if not userInput.startswith('/file '):
+        userInput = 'm' + userInput
+
+
 
     #############################
     #     FILE TRANSMISSION     #
@@ -2220,6 +2294,10 @@ while True:
     if userInput.startswith('/file '):
 
         fileName = userInput.split(' ')[1]
+        if not os.path.isfile(fileName):
+            print '\nError: Could not find file' + fileName + '\n'
+            continue
+
 
         if nick in groupFileList:
             sendTarget = 'all members of group \'' + nick + '\''
@@ -2232,21 +2310,21 @@ while True:
 
             with open('TFCtmpFile', 'r') as file:
                 b64File     = file.readlines()
-                fileMessage = ''
+                fileContent = ''
                 for line in b64File:
-                    fileMessage += line
+                    fileContent += line
 
-
-            if fileMessage == '':
+            if fileContent == '':
                 os.system('clear')
-                print 'Error: No file \'' + fileName + '\' found. Transmission aborted.\n'
+                print 'Error: target file \'' + fileName + '\'  was empty. Transmission aborted.\n'
+                continue
 
-
-            userInput = 'TFCFILE' + fileMessage
+            # All packets to recipients have header {s,l,a,e}{m,f}. 'f' stands for file.
+            userInput = 'f' + fileContent
             subprocess.Popen('shred -n ' + str(shredIterations) + ' -z -u TFCtmpFile', shell=True).wait()
 
             os.system('clear')
-            print '\nSending file ' + fileName + '\n'
+            print '\nSending file ' + fileName
 
         else:
             print '\nFile sending aborted\n'
@@ -2270,7 +2348,10 @@ while True:
 
         if not groupMemberList:
             os.system('clear')
-            print 'Current group is empty. No message was sent.\n'
+            if userInput.startswith('f'):
+                print 'Currently selected group is empty. No file was sent.\n'
+            if userInput.startswith('m'):
+                print 'Currently selected group is empty. No message was sent.\n'
             continue
 
         for member in groupMemberList:
@@ -2283,8 +2364,7 @@ while True:
             else:
                 short_msg_process(userInput, xmpp)
 
-        sleep(0.1)
-
+        time.sleep(0.1)
         print ''
